docker/requirements.txt
dateparser
pandas
datetime
azure-storage-file-datalake
bs4
requests